// frontend/src/app/gen/model/Comment.ts
/**
 * CloudRef API
 * REST API of the CloudRef system.
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

export interface Comment {
    id?: number;

    bibtexkey?: string;

    author?: string;

    publish?: boolean;

    content?: string;

    page?: string;

    pageNumber?: number;

    date?: Date;

    alterationDate?: Date;

}